// frontend/src/constants/enum/transaction.enum.ts
export enum TransactionTypeEnum {
  Income = "Income",
  Expense = "Expense",
}

export enum TransactionTagEnum {
  FoodandDrinks = "Food & Drinks",
  Shopping = "Shopping",
  Housing = "Housing",
  Transportation = "Transportation",
  Vehicle = "Vehicle",
  LifeAndEntertainment = "Life & Entertainment",
  MobileandPC = "Mobile & PC",
  FinancialExpenses = "Financial Expenses",
  Investments = "Investments",
  Income = "Income",
  Supermarket = "Supermarket",
}